requirements.txt
# Core API
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4

# Database - SQLite (no PostgreSQL needed)
sqlalchemy==2.0.30

# Image processing - lightweight
pillow==10.3.0
imagehash==4.3.1
requests==2.31.0

# Search APIs (all optional - system works without them)
python-dotenv==1.0.1
